+++ Common/Meta.cs
@@ -2,7 +2,7 @@ namespace Dyalect;
 
 internal static class Meta
 {
-    public static readonly DateTime Epoch = new(2019, 4, 17);
+    public static readonly System.DateTime Epoch = new(2019, 4, 17);
 
     public const string Version = "0.47.1";