+++ Dyalect.Library/IO/Extensions.cs
@@ -11,9 +11,9 @@ internal static class Extensions
         {
             return action();
         }
-        catch (Exception)
+        catch (Exception ex)
         {
-            ctx.IOFailed();
+            ctx.IOFailed(ex.Message);
             return default;
         }
     }
@@ -24,9 +24,9 @@ internal static class Extensions
         {
             action();
         }
-        catch (Exception)
+        catch (Exception ex)
         {
-            ctx.IOFailed();
+            ctx.IOFailed(ex.Message);
         }
     }
 }