+++ Docs/Changes.md
@@ -1,5 +1,5 @@
 # 0.47.1
-  *
+  * Change: _`IOFailed` exception should include details about an error_ ([Issue #949](https://github.com/vorov2/dyalect/issues/949)).
 
 # 0.47.0
   * _Project migrated to .NET 7.0_